requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
onnxruntime==1.22.1
modelscope==1.29.1
torch==2.7.1
torchaudio==2.7.1
python-dotenv==1.1.1


matcha-tts
openai-whisper
tensorrt==10.12.0.36
vllm==0.10.1.1
transformers
hyperpyyaml
deepspeed
pyarrow
pyworld
wetext
peft